Fixed some issues with tiff::FileWriter.  There was a memory leak of the tiff::ImageWriter pointers - addImage() allocated them but close() (and hence the destructor as well) just cleared the vector (it didn't delete them).
Changed the destructor so that it only closes the underlying file output stream if it's still open (previously it always closed it which led to a warning in valgrind about a close(-1)).

Made the class noncopyable since the compiler-generated copy constructor and assignment operator are invalid and it shouldn't be necessary to copy an object of this type.

git-svn-id: http://svn.code.sf.net/p/coda-oss/code/trunk@261 1dbd42e5-0d20-4692-8756-183d5fc073b5