Previously, the PLUGIN_EXPOSE_IDENT macros declared a static object of the desired plugin type and then returned a raw pointer to it by taking its address.  BasicPluginManager then held onto this raw pointer.  The problem with this approach was that the plugin managers that were declared as singletons could not be destructed (i.e. they needed to be instantiated with the template parameter for auto destruct set to false).  Otherwise, the static plugin objects would go out of scope prior to the singletons (because the singletons were destructed via atexit()), so the raw pointers they were holding were now pointing to invalid memory.
To get around this, the PLUGIN_EXPOSE_IDENT macros now declare static SharedPtr's and return a SharedPtr.  BasicPluginManager now also holds onto a SharedPtr.  This way, when the static SharedPtr goes out of scope, its reference count will be decremented but the BasicPluginManager's copy will still be a valid pointer.  This allows us to auto-destruct the plugin manager singletons.

git-svn-id: http://svn.code.sf.net/p/coda-oss/code/trunk@290 1dbd42e5-0d20-4692-8756-183d5fc073b5