Introduced Polygon2D
However in the function bool pointInPolygon(const Line2D::Point& p) const, there is a need to compare the points' rows and cols (which are doubles in this case). The RowCol object's ==operator is not built to compare when template to doubles (see RowCol.h), so I created a margin of error (#define MOE) for this comparison. I'm not quite sure yet what would be a good MOE or if this is sufficient--maybe epsilon would be better?