Initial partial implementation of CMake build support for CODA-oss
As of this commit, 13 modules are working, 4 are partially working, and
four are not working, mainly because the drivers implementation is TBD.

However, only seven modules are enabled in this commit, to shorten
testing times.  Also, this has only been tested on Windows.

There are also a few additions to the .gitignore, to ignore the python
cache, install directory, waf expansions, and VS local settings.