Throw helpful exception on null blockingInfo, in ImageReader.cpp
When errors occur in nitf_ImageIO_getBlockingInfo, it returns a NULL pointer and sets error to a helpful message. Previously the code wasn't checking for this, leading to a later less helpful error like "Invalid handle". This enables the more precise errors to be passed up the call stack to the caller.